backend/requirements.txt
# Backend FastAPI - Requirements
# Python 3.10+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# YouTube API
google-api-python-client==2.115.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Utility
isodate==0.6.1
python-dotenv==1.0.1
python-multipart==0.0.6

# CORS
fastapi-cors==0.0.6
